@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 // This is our custom page
 class FadePage extends Page {
-  FadePage({@required this.child, LocalKey key}) : super(key: key);
+  FadePage({required this.child, LocalKey? key}) : super(key: key);
   final Widget child;
 
   @override
