@@ -8,7 +8,7 @@ import '../fade/index.dart';
 import '../nested/index.dart';
 
 class HomeScreen extends StatelessWidget {
-  HomeScreen({Key key}) : super(key: key);
+  HomeScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -25,19 +25,19 @@ class HomeScreen extends StatelessWidget {
                   title: const Text('Books'),
                   onTap: () {
                     BaseRouterDelegate.of(context)
-                        ?.push(BooksListNavigate.path.toUri());
+                        ?.push(BooksListNavigate.path.toUri()!);
                   }),
               ListTile(
                   title: const Text('Fade'),
                   onTap: () {
                     BaseRouterDelegate.of(context)
-                        ?.push(FadeNavigate.path.toUri());
+                        ?.push(FadeNavigate.path.toUri()!);
                   }),
               ListTile(
                   title: const Text('Nested'),
                   onTap: () {
                     BaseRouterDelegate.of(context)
-                        ?.push(NestedNavigate.path.toUri());
+                        ?.push(NestedNavigate.path.toUri()!);
                   }),
             ],
           ),
