@@ -9,9 +9,9 @@ import 'index.dart';
 class FadeNavigate {
   static String path = 'fade';
   static RouteConfig routeConfig = RouteConfig(
-    state: (Uri uri) => RouteState(uri: path.toUri()),
+    state: (Uri? uri) => RouteState(uri: path.toUri()),
     isThisPage: (RouteState state) {
-      if (state?.firstPath == path) {
+      if (state.firstPath == path) {
         return true;
       }
       return false;
@@ -19,7 +19,7 @@ class FadeNavigate {
     settings: (RouteState state) {
       return null;
     },
-    page: ({RouteState state}) {
+    page: ({RouteState? state}) {
       return FadePage(key: const ValueKey('FadePage'), child: FadeScreen());
     },
   );
