@@ -10,9 +10,9 @@ class NestedNavigate {
   static String path = 'nested';
   static NavmeRouterDelegate routerDelegate = NavmeRouterDelegate.nested();
   static RouteConfig routeConfig = RouteConfig(
-    state: (Uri uri) => RouteState(uri: path.toUri()),
+    state: (Uri? uri) => RouteState(uri: path.toUri()),
     isThisPage: (RouteState state) {
-      if (state?.firstPath == path) {
+      if (state.firstPath == path) {
         return true;
       }
       return false;
@@ -20,7 +20,7 @@ class NestedNavigate {
     settings: (RouteState state) {
       return null;
     },
-    page: ({RouteState state}) {
+    page: ({RouteState? state}) {
       return MaterialPage(
           key: const ValueKey('NestedPage'),
           child: NestedScreen(
